Restructured the imports in simplify with related changes
